+++ apis/WishlistApi.ts
@@ -32,10 +32,11 @@ export class WishlistApi extends BaseWishlistApi {
     try {
       const locale = await this.getCommercetoolsLocal();
       const response = await this.getApiForProject()
+        .inStoreKeyWithStoreKeyValue({ storeKey })
         .shoppingLists()
         .get({
           queryArgs: {
-            where: [`customer(id="${accountId}")`, `store(key="${storeKey}")`],
+            where: [`customer(id="${accountId}")`],
             expand: expandVariants,
           },
         })
@@ -101,8 +102,9 @@ export class WishlistApi extends BaseWishlistApi {
   create = async (accountId: string, storeKey: string, wishlist: WishlistDraft) => {
     try {
       const locale = await this.getCommercetoolsLocal();
-      const body = WishlistMapper.wishlistToCommercetoolsShoppingListDraft(wishlist, locale, accountId, storeKey);
+      const body = WishlistMapper.wishlistToCommercetoolsShoppingListDraft(wishlist, locale, accountId);
       const response = await this.getApiForProject()
+        .inStoreKeyWithStoreKeyValue({ storeKey })
         .shoppingLists()
         .post({
           body: body,
@@ -115,6 +117,53 @@ export class WishlistApi extends BaseWishlistApi {
       return WishlistMapper.commercetoolsShoppingListToWishlist(response.body, locale);
     } catch (error) {
       throw new Error(`Create wishlist failed: ${error}`);
+    }
+  };
+
+  delete = async (wishlist: Wishlist, storeKey: string) => {
+    try {
+      await this.getApiForProject()
+        .inStoreKeyWithStoreKeyValue({ storeKey })
+        .shoppingLists()
+        .withId({ ID: wishlist.wishlistId })
+        .delete({
+          queryArgs: {
+            version: +wishlist.wishlistVersion,
+          },
+        })
+        .execute();
+    } catch (error) {
+      throw new Error(`Delete wishlist failed: ${error}`);
+    }
+  };
+
+  rename = async (wishlist: Wishlist, name: string) => {
+    const locale = await this.getCommercetoolsLocal();
+
+    try {
+      const response = await this.getApiForProject()
+        .shoppingLists()
+        .withId({ ID: wishlist.wishlistId })
+        .post({
+          body: {
+            version: +wishlist.wishlistVersion,
+            actions: [
+              {
+                action: 'changeName',
+                name: {
+                  [locale.language]: name,
+                },
+              },
+            ],
+          },
+          queryArgs: {
+            expand: expandVariants,
+          },
+        })
+        .execute();
+      return WishlistMapper.commercetoolsShoppingListToWishlist(response.body, locale);
+    } catch (error) {
+      throw new Error(`Rename wishlist failed: ${error}`);
     }
   };